Move collision handling to Cell class
