+++ Assets/Prefabs/virusnonAttached.prefab
@@ -1079,3 +1079,22 @@ MonoBehaviour:
   rpmY: 10
   rpmZ: 10
   Rotating: 1
+--- !u!114 &2826181768371610432
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5549678881499093669}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 10dc2176c7f8bcc47ba8e4849b42893d, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  type: 1
+  creationTimestamp: 0
+  targetCell: {fileID: 0}
+  sparklesPrefab: {fileID: 0}
+  explosionPrefab: {fileID: 0}
+  ObjectGenerated: {fileID: 0}
+  Antibody: {fileID: 0}
