+++ Assets/Prefabs/whiteBlood.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: be0dedad366ecd5459f91d00412109a4
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
