--- Assets/Scripts/SimulationEngine/Attach.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d668946a754a45d4b929561ab9fad2fe
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
